Frontpage from figma design (#1)
* feat: remove old css styles for redesign

* feat: add background color

change default background to #121212 according to figma design

* feat(css): red accent colour

* feat(fonts): add new fonts from figma

Adds:
- Press Start 2P font (headers)
- Space Mono font (body)

* feat: suit icons

Add the four suit icons for use on the main page and throughout the
website

* feat: main page nav fonts and icons

Update the main page navbar / header to use the new fonts and icons

* deps: add tw-merge and clsx

adds tw-merge and clsx for shadcn-ui-like classname merging on custom
components

* feat: cn util function from shadcn

helps to create tailwind-compatible custom components

* feat: generic suit icon component

* feat: rework searchbar

Rework the searchbar component with the new colours, styles and fonts.

* Revert "feat: rework searchbar"

This reverts commit d1972b1203abc0dbd2ec5c7b96973fa087a36455.

* Reapply "feat: rework searchbar"

This reverts commit c8c1a5f767929ee5b31fe83065e28602607d6199.

* feat: top games

add a top games list with new styling

* feat: update main page with search and top games

* refac: remove unecessary portal definition

* fix: LITERALLY FIRE ME

I made the card backgrounds transparent and couldn't figure out why the
"CSS Stacking Context" was so difficult to understand. THE BACKGROUND
WAS CLEAR. ARSCHLOCH

* style: fix search term hover colour

bg-gray-100 -> bg-card-shadow for consistency